added dataSource parameter to customer route
