+++ apps/single-view/src/Gateways/Person.tsx
@@ -1,17 +1,28 @@
 import axios from "axios";
 import { Person } from "../Interfaces/personInterfaces";
 import { getToken } from "../Utils/getHackneyToken";
+import { getCookie } from "../Utils/getCookie";
 
-export const getPerson = async (id: string): Promise<Person | null> => {
+export const getPerson = async (
+  id: string,
+  dataSource: number
+): Promise<Person | null> => {
   try {
-    const response = await axios.get(
-      `${process.env.PERSON_API_V1}/persons/${id}`,
-      {
+    let response;
+
+    if (dataSource == 0) {
+      response = await axios.get(`${process.env.SV_API_V1}/customers/${id}`, {
         headers: {
           Authorization: `${getToken()}`,
         },
-      }
-    );
+      });
+    } else {
+      response = await axios.get(
+        `${
+          process.env.SV_API_V1
+        }/getJigsawCustomer?id=${id}&redisId=${getCookie("jigsawToken")}`
+      );
+    }
 
     if (response.status != 200) {
       throw new Error("Error retrieving person");
