+++ apps/single-view/src/Views/SearchView/searchByResident.tsx
@@ -1,25 +1,11 @@
 import React, { useState } from "react";
 import { SearchResident } from "../../Gateways/SearchResident";
 import { housingSearchPerson } from "../../Interfaces";
+import { getCookie } from "../../Utils/getCookie";
 import { Input } from "../../Components";
 
 interface myProps {
   setResultsFunction: (searchResults: housingSearchPerson[]) => void;
-}
-
-function getCookie(cname: string) {
-  let name = cname + "=";
-  let ca = document.cookie.split(";");
-  for (let i = 0; i < ca.length; i++) {
-    let c = ca[i];
-    while (c.charAt(0) == " ") {
-      c = c.substring(1);
-    }
-    if (c.indexOf(name) == 0) {
-      return c.substring(name.length, c.length);
-    }
-  }
-  return "";
 }
 
 export const SearchByResident = (props: myProps): JSX.Element => {
