+++ apps/single-view/src/Views/SearchView/searchResults.tsx
@@ -46,7 +46,7 @@ export const SearchResults = (props: myProps): JSX.Element => {
             return (
               <div className="lbh-body" key={index}>
                 <a
-                  href={`/customers/${person.id}`}
+                  href={`/customers/${person.id}/${person.dataSource}`}
                   className="lbh-link lbh-link--no-visited-state"
                 >
                   {person.firstName} {person.surName}
